+++ frontend/src/mythicaldash/Pterodactyl/Servers.ts
@@ -7,6 +7,14 @@ class Servers {
         return data.servers;
     }
 
+    public static async getPterodactylQueuedServers() {
+        const response = await fetch('/api/user/session/servers', {
+            method: 'GET',
+        });
+        const data = await response.json();
+        return data.servers_queue;
+    }
+
     public static async getPterodactylResources() {
         const response = await fetch('/api/user/session/pterodactyl/resources', {
             method: 'GET',